+++ src/constants.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sun May  2 00:01:39 2021
+
+@author: oriol
+"""
+
+NO_GPS_TEXT = 'No GPS signal'
+SEARCHING_GPS_TEXT = '...searching GPS...'
+NO_GPS_NUM = 1000
+SEARCHING_GPS_NUM = 2000