+++ docs/changelog.md
@@ -4,4 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). For the version numbers, I just use a simple 2-digit, for major and minor changes. Each version has its corresponding apk under the folder `releases`.
 
-## Unreleased
+
+## 0.1 - 2021-05-02
+
+## Added
+- GPS activation (permissions, etc.), show current GPS coordinates in the screen
+- Button to switch on/off GPS
+- Status icon and text, showing if GPS is on, searching or off
+- Accuracy icon to show the level of accuracy of the coordinates